requirements.txt
# Core Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
bcrypt>=4.0.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# File Processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# QR Code & Barcode Generation
qrcode[pil]>=7.0.0
python-barcode[images]>=0.15.0
Pillow>=10.0.0

# PDF Generation
reportlab>=4.0.0

# Logging
loguru>=0.7.0

# System Monitoring
psutil>=5.9.0